Backend/requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
joblib==1.0.1
nltk==3.5
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
tqdm==4.59.0
